requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
line-bot-sdk==2.3.0